Mejoras en accesibilidad y rendimiento
