+++ package.json
@@ -24,7 +24,7 @@
     "@angular/platform-server": "^19.2.0",
     "@angular/router": "^19.2.0",
     "@angular/ssr": "^19.2.4",
-    "@ngx-translate/core": "^16.0.3",
+    "@ngx-translate/core": "^16.0.4",
     "@ngx-translate/http-loader": "^16.0.1",
     "express": "^4.18.2",
     "ngx-toastr": "^17.0.0",
