+++ src/app/ajustes/ajustes.component.html
@@ -13,7 +13,7 @@
           <div style="height: 87px;">
             <div><label>{{ 'AJUSTES.PAIS' | translate }}</label></div>
             <div class="category-selector">
-              <select class="category-dropdown" formControlName="pais">
+              <select class="category-dropdown" formControlName="pais" title="{{ 'AJUSTES.SELECCIONAR_PAIS' | translate }}">
                 <option value="" disabled>{{ 'AJUSTES.SELECCIONAR_PAIS' | translate }}</option>
                 <option value="col">Colombia</option>
                 <option value="arg">Argentina</option>
