+++ src/app/barra-superior/barra-superior.component.html
@@ -1,5 +1,5 @@
 <div class="topbar">
-  <img src="assets/LOGO-CCP.png" alt="Logo CCP" class="logo">
+  <img loading="lazy" src="assets/LOGO-CCP.png" alt="Logo CCP" class="logo">
   <div class="search-container">
     <form [formGroup]="searchForm" (ngSubmit)="onSubmit()">
       <input 
@@ -12,7 +12,7 @@
     
   </div>
   <div class="topbar-icons">
-    <img src="assets/bell.svg" alt="Notificaciones" class="icon">
-    <img src="assets/user.png" alt="Usuario" class="user-avatar">
+    <img loading="lazy" src="assets/bell.svg" alt="Notificaciones" class="icon">
+    <img loading="lazy" src="assets/user.png" alt="Usuario" class="user-avatar">
   </div>
 </div>
