+++ src/app/bienvenida/bienvenida.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { BarraSuperiorComponent } from '../barra-superior/barra-superior.component';
 import { MenuLateralComponent } from '../menu-lateral/menu-lateral.component';
-import { TranslateModule } from '@ngx-translate/core';
 import { I18nModule } from '../i18n.module';
 
 @Component({
@@ -10,8 +9,7 @@ import { I18nModule } from '../i18n.module';
   imports: [
     BarraSuperiorComponent,
     MenuLateralComponent,
-    I18nModule,
-    TranslateModule
+    I18nModule
   ],
   templateUrl: './bienvenida.component.html',
   styleUrls: ['./bienvenida.component.css']
