+++ src/app/i18n.module.ts
@@ -1,25 +1,1 @@
-import { NgModule } from '@angular/core';
-import { HttpClientModule, HttpClient } from '@angular/common/http';
-import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
-import { TranslateHttpLoader } from '@ngx-translate/http-loader';
-
-export function HttpLoaderFactory(http: HttpClient) {
-  return new TranslateHttpLoader(http, './assets/i18n/', '.json');
-}
-
-@NgModule({
-  imports: [
-    HttpClientModule,
-    TranslateModule.forRoot({
-      //defaultLanguage: 'es',
-      useDefaultLang: false,
-      loader: {
-        provide: TranslateLoader,
-        useFactory: HttpLoaderFactory,
-        deps: [HttpClient]
-      }
-    })
-  ],
-  exports: [TranslateModule]
-})
-export class I18nModule {}
+export { TranslateModule as I18nModule } from '@ngx-translate/core';
