+++ src/app/menu-lateral/menu-lateral.component.css
@@ -62,7 +62,7 @@
     margin-top: -1.00px;
     font-family: "Plus Jakarta Sans-Medium", Helvetica;
     font-weight: 500;
-    color: #a1a1aa;
+    color: #5a5a5a;
     font-size: 11px;
     letter-spacing: 1.00px;
     line-height: 18px;
@@ -106,7 +106,7 @@
     left: 2px;
     font-family: "Plus Jakarta Sans-Medium", Helvetica;
     font-weight: 500;
-    color: #a1a1aa;
+    color: #5a5a5a;
     font-size: 11px;
     letter-spacing: 1.00px;
     line-height: 18px;
