+++ src/app/products/product-registration/product-registration.component.spec.ts
@@ -3,7 +3,8 @@ import { ProductRegistrationComponent } from './product-registration.component';
 import { ReactiveFormsModule } from '@angular/forms';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { ToastrService } from 'ngx-toastr';
-import { of, throwError } from 'rxjs';
+import { of } from 'rxjs/internal/observable/of';
+import { throwError } from 'rxjs/internal/observable/throwError';
 import { ManufacturerService } from '../../core/services/manufacturer.service';
 
 describe('ProductRegistrationComponent', () => {
