+++ src/app/products/registration-type/registration-type.component.ts
@@ -1,8 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 import { BarraSuperiorComponent } from '../../barra-superior/barra-superior.component';
 import { MenuLateralComponent } from '../../menu-lateral/menu-lateral.component';
 import { Router } from '@angular/router';
-import { TranslateModule } from '@ngx-translate/core';
 import { I18nModule } from '../../i18n.module';
 
 @Component({
@@ -11,7 +10,6 @@ import { I18nModule } from '../../i18n.module';
   imports: [
     MenuLateralComponent,
     BarraSuperiorComponent,
-    TranslateModule,
     I18nModule
   ],
   templateUrl: './registration-type.component.html',
