+++ src/app/sellers/seller-registration/seller-registration.component.html
@@ -50,7 +50,7 @@
                 <label for="country">{{ 'VENDEDOR.PAIS' | translate }}</label>
               </div>
               <div class="category-selector">
-                <select class="category-dropdown" formControlName="country">
+                <select class="category-dropdown" formControlName="country" title="{{ 'VENDEDOR.SELECCIONAR_PAIS' | translate }}">
                   <option value="" disabled>{{ 'VENDEDOR.SELECCIONAR_PAIS' | translate }}</option>
                   <option value="Argentina">Argentina</option>
                   <option value="Brasil">Brasil</option>
