+++ src/app/sellers/seller-registration/seller-registration.component.ts
@@ -1,12 +1,11 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 import { BarraSuperiorComponent } from '../../barra-superior/barra-superior.component';
 import { MenuLateralComponent } from '../../menu-lateral/menu-lateral.component';
 import { FormBuilder, Validators, ReactiveFormsModule, FormGroup } from '@angular/forms';
 import { CommonModule } from '@angular/common';
-import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 import { ToastrService } from 'ngx-toastr';
 import { environment } from '../../../environments/environment';
-import { TranslateModule } from '@ngx-translate/core';
 import { I18nModule } from '../../i18n.module';
 
 @Component({
@@ -17,8 +16,7 @@ import { I18nModule } from '../../i18n.module';
     MenuLateralComponent,
     ReactiveFormsModule,
     CommonModule,
-    I18nModule,
-    TranslateModule
+    I18nModule
   ],
   templateUrl: './seller-registration.component.html',
   styleUrls: ['./seller-registration.component.css']
