+++ src/app/supplier/supplier-registration/supplier-registration.component.html
@@ -33,7 +33,7 @@
                 <label for="country">{{ 'REGISTRO_PROVEEDORES.PAIS' | translate }}</label>
               </div>
               <div class="category-selector">
-                <select class="category-dropdown" formControlName="country">
+                <select class="category-dropdown" formControlName="country" title="{{ 'REGISTRO_PROVEEDORES.SELECCIONAR_PAIS' | translate }}">
                   <option value="" disabled>{{ 'REGISTRO_PROVEEDORES.SELECCIONAR_PAIS' | translate }}</option>
                   <option value="Argentina">Argentina</option>
                   <option value="Brasil">Brasil</option>
