+++ src/index.html
@@ -5,7 +5,7 @@
   <title>CCP</title>
   <base href="/">
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;600;700&display=swap" rel="stylesheet">
+  <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;600;700&display=swap" rel="preload" as="style">
 </head>
 <body>
   <app-root></app-root>
